Added enemy spawning to the game over screen. - since the player is dead but still exists in the run time as an object with a position, collision with enemies is still possible so in the collision handler for enemy to player collisions we have to put in a check to see if we are in a game over state or not and if we are then ignore the collision - need to clear all enemies a few times and from different classes so I made a clearenemies method in the enitity manager - clear all the enemies in the playership class when the player presses a button to start a new game after the game over state - added an extra OR statement to the enemy spawner to run the spawn if the game is in the game over state - also had to remove spawning of seekers when in the game over state since the seekers behaviour requires a player to follow

+++ GeoWar.Shared/EntityManager.cs
@@ -87,6 +87,13 @@ namespace GeoWar
             }
         }
 
+        public static void ClearEnemies()
+        {
+            // this is a linq function that iterates through the list and runs the was shot 
+            // method on all enemies in the enemies list to kill them off
+            enemies.ForEach(enemy => enemy.WasShot());
+        }
+
         private static bool IsColliding(Entity a, Entity b)
         {
             // find the sum of the two entities radiuses
@@ -139,13 +146,12 @@ namespace GeoWar
             for (int enemyIndex = 0; enemyIndex < enemies.Count; enemyIndex++)
             {
                 // if the enemy is active and has collided with the player
-                if (enemies[enemyIndex].IsActive && IsColliding(PlayerShip.Instance, enemies[enemyIndex]))
+                if (enemies[enemyIndex].IsActive && IsColliding(PlayerShip.Instance, enemies[enemyIndex]) && !PlayerStatus.IsGameOver)
                 {
                     // run the kill method on the player to kill him
                     PlayerShip.Instance.Kill();
-                    // this is a linq function that iterates through the list and runs the was shot 
-                    // method on all enemies in the enemies list to kill them off once the player dies
-                    enemies.ForEach(enemy => enemy.WasShot());
+                    // kill all the enemies off when the player dies
+                    ClearEnemies();
                     // immediatly exit the loop once we find that the player is now dead
                     break;
                 }
